ci(psalm): Fix remaining psalm errors

+++ lib/Service/PersonalDefaultsService.php
@@ -165,7 +165,8 @@ class PersonalDefaultsService {
 			return null;
 		}
 		for ($len = 3; $len >= 1; $len--) {
-			$prefix = substr($digits, 0, $len);
+			// Numeric-string array keys are stored as ints, so match with an int offset.
+			$prefix = (int)substr($digits, 0, $len);
 			if (isset(self::CALLING_CODES[$prefix])) {
 				return self::CALLING_CODES[$prefix];
 			}
